[BDN-503] Make bid requests in parallel (#399)
* [BDN-503] Remove unused SegmentMatcher and fix warning

* [BDN-503] Make bid requests in parallel

+++ internal/sdkapi/schema/bidding_request.go
@@ -13,7 +13,7 @@ type BiddingRequest struct {
 	Adapters Adapters `json:"adapters" validate:"required"`
 	Imp      Imp      `json:"imp" validate:"required"`
 	Test     bool     `json:"test"` // Flag indicating that request is test
-	TMax     int      `json:"tmax"` // Max response time for server before timeout
+	TMax     int64    `json:"tmax"` // Max response time for server before timeout
 }
 
 func (b *BiddingRequest) NormalizeValues() {
